% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynseg.R
\name{comp_dynseg__fast__}
\alias{comp_dynseg__fast__}
\title{Segments the data using dynamic programming algorithm}
\usage{
comp_dynseg__fast__(data, ncol = NULL, segthr = NULL)
}
\arguments{
\item{segthr}{Threshold on the number of block segments to fit the model.
This is highly recommend for perfomance issues on large datasets}
}
\description{
Uses dynamical segmentation to obtain the
log likelihood, changepoints and probabilities
as a function of the number of segments
}
