% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{compare_model}
\alias{compare_model}
\title{Compare or evaluate model performance with respect
to other model or ground truth}
\usage{
compare_model(model1, model2, m = NULL)
}
\arguments{
\item{model1}{The first bincpd object or list of sorted integers
representing the change point set.}

\item{model2}{The second bincpd object or list of sorted integers
representing the change point set.}

\item{m}{The number of variables which the model was fitted on. Only needs to
be passed if both arguments are change point sets instead of a bincpd object.}
}
\description{

}
