% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cvseg.R
\name{compute_cvseg}
\alias{compute_cvseg}
\title{Segments the data using dynamical programming with cross validation on the
number of segments}
\usage{
compute_cvseg(data_matrix, segthr = NULL, n_folds = 5)
}
\arguments{
\item{data_matrix}{Data matrix to perform change point analysis}

\item{segthr}{Threshold on the number of block segments to fit the model.
This is highly recommend for performance issues on large data sets}

\item{n_folds}{Number of folds used in the k-fold CV}
}
\description{
Minimizes the k-fold CV mean negative log likelihood computed, no
penalization term included. For each fold, the dynseg algorithm is run and we
compute the best estimator best estimator for each number of change points
and their respective log likelihoods. In the end, the algorithm is rerun in
the whole data set using and select the best estimator from the best number
of change points.
}
