% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dynseg.R
\name{compute_dynseg}
\alias{compute_dynseg}
\title{Segments the data using dynamic programming algorithm}
\usage{
compute_dynseg(data, segthr = NULL, lambda = 1, pen_func = bic_loss_hs)
}
\arguments{
\item{segthr}{Threshold on the number of block segments to fit the model.
This is highly recommend for perfomance issues on large datasets}
}
\description{
Uses dynamical segmentation to obtain the
log likelihood, changepoints and probabilities
as a function of the number of segments
}
