% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fusedlasso.R
\name{compute_fusedlasso}
\alias{compute_fusedlasso}
\title{Computes the Fused Lasso estimator.}
\usage{
compute_fusedlasso(
  data_matrix,
  lambda = 1,
  sample_pen_func = log,
  precision = 0.001,
  MLE_prob = FALSE,
  check_dcp = FALSE
)
}
\arguments{
\item{data_matrix}{Data matrix to perform change point analysis}

\item{lambda}{Penalization constant}

\item{sample_pen_func}{The penalization function J(n) multiplying
the fused regularization term}

\item{precision}{Precision to calculate the change points.
If \eqn{|p_{i+1} - p_{i}| > precision}, then \eqn{i} is considered
a change point.}

\item{MLE_prob}{If TRUE, the probability parameter will be estimated using
Maximum likehood approach, which is the mean of the block. So only
the change points will be determined by the L1 loss.}

\item{check_dcp}{a flag to check if the optimization problem satisfies the
DCP rules. Checking the rules is expensive, and since our problem satisfies
the rules, the default is FALSE.}
}
\description{
The estimator is parameter that minimizes the fused lasso loss defined by
\eqn{-l(p; X) + J(n) \times lambda \times sum_{i=1}^m|p_i - p_{i+1}|},
where l is the log likelihood, and the penalization is on the difference of
the probability parameters
}
