% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_segmentation.R
\name{compute_hierseg}
\alias{compute_hierseg}
\title{Segments the data using hierarchical algorithm}
\usage{
compute_hierseg(data_matrix, lambda = 1, pen_func = bic_loss_hs)
}
\arguments{
\item{pen_func}{A penalization function defined i integer intervals
The function signature should be pen(left_index, right_index, n, m),
where the left_index:right_index is the integer interval, n the sample
size and m the number of variables/columns.}
}
\description{
Uses binary splitting in a greedy solution approach to obtain the
log likelihood, change points and probabilities
as a function of the number of segments
}
