% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{fit_cv_pml}
\alias{fit_cv_pml}
\title{Computes the penalized maximum likelihood estimator using cross validation
Estimates the number of segments using maximum likelihood and
K-fold cross validation}
\usage{
fit_cv_pml(
  data_mat,
  dynseg = NULL,
  segthr = NULL,
  nfolds = 5,
  c_interval = seq(0.5, 10, by = 0.25),
  pen_func = NULL,
  seed = NULL,
  progress_omit = FALSE,
  time_omit = FALSE,
  auto_search = FALSE,
  search_limit = 5
)
}
\arguments{
\item{segthr}{Threshold on the number of block segments to fit the model.
This is highly recommend for perfomance issues on large datasets}

\item{c_interval}{penalization constant set. Each value of the set will be
used to compute}

\item{progress_omit}{Omit \% of progress.}

\item{time_omit}{Omit time taken to run the estimator.}

\item{n_folds}{number of k-folds in cross valition}
}
\description{
After splitting the data in a k-fold CV form, the estimator calculates
the penalized maximum likehood estimator, which is the
argmin of -l(p; X) + lambda*R(p;X), for the train data.
Then, it calculates the likelihood on the validation data. The estimator is
then recalculated using the whole dataset for the number of segmetns that
had the highest likelihood average.
}
