% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{fit_cv_seg}
\alias{fit_cv_seg}
\title{Estimates the number of segments using maximum likelihood and
K-fold cross validation}
\usage{
fit_cv_seg(
  data_mat,
  dynseg = NULL,
  segthr = NULL,
  nfolds = 5,
  seed = NULL,
  progress_omit = FALSE,
  time_omit = FALSE
)
}
\arguments{
\item{segthr}{Threshold on the number of block segments to fit the model.
This is highly recommend for perfomance issues on large datasets}

\item{progress_omit}{Omit \% of progress.}

\item{time_omit}{Omit time taken to run the estimator.}

\item{n_folds}{number of k-folds in cross valition}
}
\description{
After splitting the data in a k-fold CV form, the estimator calculates,
for each number of segments k, the maximum likehood
estimator conditional to that number of segments on the train data.
Then, it calculates the likelihood on the validation data. The estimator is
then recalculated using the whole dataset for the number of segmetns that
had the highest likelihood average.
}
