% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso.R
\name{fit_fused_cvxr}
\alias{fit_fused_cvxr}
\title{Computes the Fused Lasso estimator.}
\usage{
fit_fused_cvxr(
  data,
  lambda = 1,
  fused_prec = 1e-04,
  check_dcp = FALSE,
  MLE_prob = FALSE
)
}
\arguments{
\item{lambda}{Penalization constant}

\item{fused_prec}{Precision to calculate the change points
if |p_{i+1} - p_{i}| > fused_prec, then {i} is considered a change point.}

\item{MLE_prob}{If TRUE, the probability parameter will be estimated using
Maximum Likehood approach (in thi case, the mean of the block). So only
the breaks will be determined by the L1 loss.}
}
\description{
The estimator is the argmin of -l(p; X) + lambda*sum_{i=1}^m|p_i - p_{i+1}|,
where l is the log likelihood, and the penalization is on the difference of
the probability parameters
}
