% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynseg.R
\name{fit_pml}
\alias{fit_pml}
\title{Computes the penalized maximum likelihood estimator}
\usage{
fit_pml(data, segthr = NULL, lambda = 1, pen_func = NULL, dynseg = NULL)
}
\arguments{
\item{segthr}{Threshold on the number of block segments to fit the model.
This is highly recommend for perfomance issues on large datasets}

\item{lambda}{Penalization constant}

\item{pen_func}{The penalization function used for the computation of the estimator.
A user specified function can be provided, and the function signature is
f(n, m, prob, cp)
The default loss is the bic_loss -log(n)log(m)(len(cp)+1).}
}
\description{
The estimator is the argmin of -l(p; X) + lambda*R(p;X), where l is the
log likelihood and R is the penalization function
}
