% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{plot.bincpd}
\alias{plot.bincpd}
\title{Plot change point blocks with probabilties}
\usage{
\method{plot}{bincpd}(object, ...)
}
\description{
Method of the generic \link[=plot]{plot} for bincpd objects.
}
