% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Data prep for ROH island analysis
Description: Process SNPs data so it is ready for model fitting.
The preprocessing consists of computing the mean in a neighboorhood
of the specified radius (total size = 2*radius + 1)
and checking if it is above a cutoff (default 95\%).
If that is the case, we output 1 to that row and SNP position in a new
dataframe with the same dimensions as the original, otherwise we output 0.}
\usage{
preprocess(snps_raw, radius = 15, thr_mis_snps = 0.01, thr_mis_ind = 0.05)
}
\description{
Data prep for ROH island analysis
Description: Process SNPs data so it is ready for model fitting.
The preprocessing consists of computing the mean in a neighboorhood
of the specified radius (total size = 2*radius + 1)
and checking if it is above a cutoff (default 95\%).
If that is the case, we output 1 to that row and SNP position in a new
dataframe with the same dimensions as the original, otherwise we output 0.
}
