% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBBM.R
\name{rBBM}
\alias{rBBM}
\title{Sampler for the Binary CP Block Model}
\usage{
rBBM(
  n = 100,
  m = 1000,
  k = 10,
  prob_vec = NULL,
  changepoints = NULL,
  prob_NA = NULL
)
}
\arguments{
\item{n}{Sample size}

\item{m}{Array length}

\item{k}{Number of change points.  The number of blocks is \eqn{k + 1}.}

\item{prob_vec}{Vector \eqn{k + 1} dimensional containing the probability
parameter of each block.}

\item{changepoints}{A \eqn{k} dimensional vector containing the
change point locations. The change points are between 1 and \eqn{m-1}}

\item{prob_NA}{Probability of each entry of being NA}
}
\description{
Creates a \eqn{n \times m} matrix with \eqn{k} change points. In between
change points, the random variables are i.i.d. Bernoullis with the same
probability parameter.
}
