% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_hierseg.R
\name{compute_hierseg}
\alias{compute_hierseg}
\title{Segments the data using hierarchical algorithm}
\usage{
compute_hierseg(data_matrix, lambda = 1, pen_func = bic_loss_hs)
}
\arguments{
\item{data_matrix}{Data matrix to perform change point analysis}

\item{lambda}{Penalization constant}

\item{pen_func}{A penalization function defined i integer intervals
The function signature should be pen(left_index, right_index, n, m),
where the left_index:right_index is the integer interval, n the sample
size and m the number of variables/columns.}
}
\description{
Uses binary splitting in a greedy solution of the regularized loss
optimization problem to obtain the change points and probabilities of each
block
}
