% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fits a binary CPD model according to the method chosen}
\usage{
fit_model(data, method = "hierseg", arguments = list())
}
\description{
Fits a binary CPD model according to the method chosen
}
